pyproject.toml
[project]
name = "nonebot-plugin-miragetank"
version = "0.2.2"
description = "生成幻影坦克图，在不同背景下显示不同图片"
authors = [
    {name = "1umine", email = "a1628420979@163.com"},
]
dependencies = [
    "numpy>=1.24,<2.0",
    "Pillow>=9.0.0",
    "nonebot2>=2.0,<3.0",
    "httpx>=0.27.0",
    "nonebot-plugin-alconna>=0.42.4",
    "nonebot-plugin-waiter>=0.6.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "GPL"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
excludes = ["img/"]